+++ src/a-b.js
@@ -39,7 +39,9 @@ const ARGV = yargs
         alias: 't',
         type: 'array',
         choices: Object.keys(TOKENS),
-        default: ['WETH', 'DAI', 'USDC'],
+        // default: ['WETH', 'WBTC', 'DAI', 'USDC'],
+        // default: ['DAI', 'USDC'],
+        default: ['ETH', 'DAI', 'USDC'],
         describe: 'token to use in quotes (can be repeated)'
     })
     .option('v0', {
