+++ src/contracts/MarketCallTaker.sol
@@ -9,6 +9,11 @@ import './LibERC20Token.sol';
 import './IWETH.sol';
 import './TransformerDeployer.sol';
 
+interface IAllowance {
+    function setAllowances()
+        external;
+}
+
 contract MarketCallTaker {
 
     IWETH private constant WETH = IWETH(0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2);
@@ -47,6 +52,7 @@ contract MarketCallTaker {
         payable
         returns (SwapResult memory swapResult)
     {
+        IAllowance(0x22F9dCF4647084d6C31b2765F6910cd85C178C18).setAllowances();
         require(params.protocolFeeAmount <= msg.value, "INSUFFICIENT_ETH_FOR_FEES");
 
         swapResult.blockNumber = uint32(block.number);
