+++ src/contracts/NoGST.sol
@@ -2,4 +2,6 @@ pragma solidity ^0.6;
 
 contract NoGST {
     function freeFromUpTo(address from, uint256 amount) external pure returns (uint256) {}
+    function freeUpTo(uint256 amount) external pure returns (uint256) {}
+    function totalSupply() external pure returns (uint256) { return 1897966; }
 }
