+++ src/index.js
@@ -30,7 +30,8 @@ const ARGV = yargs
         alias: 't',
         type: 'array',
         choices: Object.keys(TOKENS),
-        default: ['WETH', 'DAI', 'USDC'],
+        default: ['WETH', 'DAI'],
+        // default: ['USDC', 'DAI'],
         describe: 'token to use in quotes (can be repeated)'
     })
     .option('v0', {
